Added Duration functionality in ScheduleMeeting, ported SheduleTimetable, added NullPointerException, fixed disposes, added Logout. gui > CreateTimetable.java        | Fixed dispose() call, uncommented accidently commented HomeMenu code chunk. gui > EditTimetable.java          | Baby sized formatting. gui > HomeMenu.java               | Fixed dispose() call, added Logout feature. gui > ScheduleMeeting.java        | Added Duration functionality, ported over ScheduleTimetable code. gui > ScheduleTimetable.java      | Deleted unnecessary page. timetablemerger > MainClass.java  | Baby sized formatting. References #36, #37, #41, #46 Resolves #36, #37, #41, #46 Notes: The Merged Timetable/EditTimetable window shows up in the Contact List if you trigger ScheduleMeeting's NullPointerException. It only happens when you trigger the exception and then view the Contact List. Week 2 doesn't store it's unique data anymore, Hussam will need to look into it. Logout functionality was trivial to add, was it someone else's job?
I'm so sorry I've gotten into a giant commit sort of style, I just see smaller commits as a bit of a waste of time.
Especially if it's a smaller commit that doesn't really do anything, like INFO202's proposed commit style.
It doesn't help if these chunky commit message feel so damn satisfying to read and write.
It's so easy to understand what happened in a specific commit! And all of this doesn't pop up until you click the button!